backend/apps/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
python-dotenv
gunicorn
django-cors-headers
